feat: use floats instead of doubles

+++ lib/src/main/java/com/github/shimeoki/jfx/rasterization/triangle/TriangleBarycentrics.java
@@ -2,11 +2,11 @@ package com.github.shimeoki.jfx.rasterization.triangle;
 
 public interface TriangleBarycentrics {
 
-    public double lambda1();
+    public float lambda1();
 
-    public double lambda2();
+    public float lambda2();
 
-    public double lambda3();
+    public float lambda3();
 
     public boolean inside();
 }
